requirements.txt
flask>=2.0.1
flask-limiter>=3.3.0
flask-caching>=2.0.2
google-generativeai>=0.3.0
python-dotenv>=0.19.0
requests>=2.26.0
gunicorn>=20.1.0
